[Modify] 댓글 추가/삭제 UUID -> userPricipal로 수정

+++ src/main/java/nova/mjs/comments/service/CommentsService.java
@@ -46,23 +46,30 @@ public class CommentsService {
 
     // 2. POST 댓글 작성
     @Transactional
-    public CommentsResponseDto.CommentSummaryDto createComment(UUID communityBoardUuid, String content, UUID memberUuid) {
-        Member member = getExistingMember(memberUuid);
+    public CommentsResponseDto.CommentSummaryDto createComment(UUID communityBoardUuid, String content, String email) {
+        // 이메일을 이용하여 현재 로그인한 회원 정보 가져오기
+        Member member = memberRepository.findByEmail(email)
+                .orElseThrow(MemberNotFoundException::new);
         CommunityBoard communityBoard = getExistingBoard(communityBoardUuid);
 
         Comments comment = Comments.create(communityBoard, member,content);
         Comments savedComment = commentsRepository.save(comment);
 
-        log.debug("댓글 작성 성공. UUID = {}", savedComment.getUuid());
+        log.debug("댓글 작성 성공. UUID = {}, 작성자 : {}", savedComment.getUuid(), email);
         return CommentsResponseDto.CommentSummaryDto.fromEntity(savedComment);
     }
 
     // 3. DELETE 댓글 삭제
     @Transactional
-    public void deleteCommentByUuid(UUID commentUuid) {
+    public void deleteCommentByUuid(UUID commentUuid, String email) {
         Comments comment = getExistingCommentByUuid(commentUuid);
+        // 현재 로그인한 사용자가 댓글 작성자인지 체크
+        if (!comment.getMember().getEmail().equals(email)) {
+            throw new IllegalArgumentException("본인이 작성한 댓글만 삭제할 수 있습니다.");
+        }
+
         commentsRepository.delete(comment);
-        log.debug("댓글 삭제 성공. ID = {}", commentUuid);
+        log.debug("댓글 삭제 성공. ID = {}, 작성자: {}", commentUuid, email);
     }
 
     // 4. 특정 게시글 존재 여부 확인
